+++ src-tauri/src/mods/valuable.rs
@@ -1,8 +1,8 @@
 use super::{redondeo, Config, Formato, Pesable, Producto, Res, Rubro};
 use serde::{Deserialize, Serialize};
+use sqlx::{Pool, Sqlite};
 use std::fmt::{self, Display};
 use Valuable as V;
-use sqlx::{Pool, Sqlite};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum Valuable {
     Prod((u8, Producto)),
@@ -148,15 +148,15 @@ impl Presentacion {
             Presentacion::Kg(_) => String::from("Kg"),
         }
     }
-    pub fn build(presentacion: &str, cantidad: f64) -> Presentacion {
+    pub fn build(presentacion: &str, cantidad: f32) -> Presentacion {
         match presentacion {
-            "Gr" => Presentacion::Gr(cantidad as f32),
+            "Gr" => Presentacion::Gr(cantidad),
             "Un" => Presentacion::Un(cantidad as u16),
-            "Lt" => Presentacion::Lt(cantidad as f32),
+            "Lt" => Presentacion::Lt(cantidad),
             "Ml" => Presentacion::Ml(cantidad as u16),
             "CC" => Presentacion::CC(cantidad as u16),
-            "Kg" => Presentacion::Kg(cantidad as f32),
-            a => panic!("No existe {}",a)
+            "Kg" => Presentacion::Kg(cantidad),
+            a => panic!("No existe {}", a),
         }
     }
 }
